feat: improve access to the GDAL cli
BREAKING CHANGE: this changes how to get access to a new gdal instance to Gdal.create()
