+++ packages/cli/src/index.ts
@@ -1,4 +1,5 @@
 export { CogBuilder } from './cog/builder';
-export { GdalCogBuilder } from './gdal/gdal';
+export { GdalCogBuilder } from './gdal/gdal.cog';
 export { GdalCogBuilderOptions } from './gdal/gdal.config';
 export { CogJobFactory } from './cog/job';
+export { Gdal } from './gdal/gdal';
